requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
python-multipart
pgvector
sqlalchemy
psycopg[binary]
pypdf
sentence-transformers
transformers
google
google-generativeai
typer
streamlit
bs4